Clock updates are now mostly correct
It is now possible to align the first positive
edge of a series of clocks. Additionally, clocks
are able to compute their current value and next
edge correctly.

I've also changed the calling behavior of eval()
to try to do it as often as possible in the
generic testbench header. This makes it less
likely to be abused by derived classes.