async_fifo: fix ram-collision bug
The underlying RAM module prevents read-write
collisions but the FIFO previously did not
deassert the read-enable and write-enable signals
during these collisions. The result of this was to
prevent the RAM (and consequently FIFO) from
allowing any information to be read during
constant, simultaneous reading and writing.

This commit also introduces a cocotb test for
constant simultaneous reads/writes.