fir_poly: add input pipeline stage
This adds a few extra registers as
overhead. However, it may be possible to remove
them. Additionally, it has several benefits. It is
easier to understand than the previous version. It
also holds input data stable for each bank for the
duration of the 2MHz period, which makes it easier
to get the correct data to the time-multiplexed
multiplier. For these reasons, it should make it
easier to make adjustments in the future when
needed.